Reduce CPU pressure around BufferLeakDetectionTest
Motivation:
Running buffer tests in parallel reduces the CPU resources available to each individual test, causing tests that rely on timeliness to suffer as a result.

Modification:
Run the BufferLeakDetectionTest in isolation from other tests.
This test demands a lot of CPU resources on its own, as it stresses the JVM garbage collector.
It also expects many tasks to complete within 20 seconds, which may not be feasible on CI, if all the other buffer tests are running.

Result:
Hopefully this will let it complete in time, and make it pass as often as before.