Add method to get read-only Buffer copy without offsets
Motivation:
It seems to be the common case that when we want a read-only copy of a buffer, we want a copy of the readable byte range.

Modification:
Add a method specifically for this use case, where we can request a read-only copy, without having to specify the readable byte range explicitly.
Also update every relevant usage site.

Result:
Cleaner code, and easier access to the structural-sharing copy optimisation.