Add a BufferAllocator.copyOf(String, Charset)
Motivation:
It is somewhat common to allocate a Buffer from a String and a Charset.

Modification:
Add a convenience method that converts the string to bytes, and allocates a buffer for those bytes, in one go.
In certain cases, we can also optimise a bit, and use the byte array we get from String as the backing memory of the Buffer.
Many usages sites have also been updated to take advantage of this new API.

Result:
Cleaner code.