Fix data race in `Http2ConnectionRoundtripTest`
Motivation:
With HTTP/2 preface flushing from #12420, we need to wait for an additional message before we can be sure that we've received the headers frame.

Modification:
Increment the request count down latch by one in `Http2ConnectionRoundtripTest.priorityUsingHigherValuedStreamIdDoesNotPreventUsingLowerStreamId`.
This increment was already done in multiple other tests in #12420, but not in this one.
Usually that was "fine", but it left the headers check in the test racing with the headers frame.
This made the test flaky.

Result:
Now the test is always passing.