Remove the last references to JUnit 4 (#12442)
Motivation:
This completes our migration to JUnit 5.

Modification:
Remove all JUnit 4 dependencies.
Replace or remove the last references to JUnit 4 types in the code.
Add a revapi exception because a public TestUtils method was exposing JUnit 4 types and had to be removed.

Result:
No more JUnit 4 in our project.

Fixes #10757