Check and skip if post body parameter name and value are empty
Motivation:

HTTP form bodies sometimes have RFC illegal extra ampersands in them; this has become accepted by lots of frameworks and so it makes it difficult to migrate to Netty.

This builds on top of #12016

Contextual information:
A POST body with the following was not accepted by Netty:
&foo=bar&baz=bang
Now it will be.

Modification:

Skips name and valueless attributes silently

Result:

Library now compatible with the broken internets