Bulk write and copyInto fixes
Motivation:
The different buffer implementations should behave the same as much as possible.
Some bulk-write methods and copyInto methods did not behave the same with regards to error handling and what exceptions were thrown.
Additionally, it also needs to be possible to perform writes to a buffer that is shared, such as when components are being iterated.

Modification:
Increase test coverage of writeBytes and copyInto methods.
Align all of their thrown exceptions.
The writeBytes method now only call ensureWritable if expansion is needed, and now respects the implicit capacity limit as it should.

Result:
More consistent behaviours from bulk methods on Buffer.