Make SensitiveBufferAllocator respect memory manager override
Motivation:
We need to override the memory manager used by the SensitiveBufferAllocator, so we can make sure we cover all implementations.
The SensitiveBufferAllocator was capturing the memory manager on allocation, but our tests only override when the allocator is obtained.

Modification:
If a MemoryManager override is set when the SensitiveBufferAllocator is obtained, then we now capture it and use it when allocating, unless the allocation site has a another override.

Result:
We now cover all possible buffer implementations in our tests, with the SensitiveBufferAllocator.