Avoid exponential back-off in PemReader
Motivation:
The key and certificate patterns used in PemReader were susceptible to exponential back-off.
The reason is that they had two adjacent matching groups that overlapped in their matching characters.
For instance, it is not clear to the matcher if a long string of newlines after the header actually belong to the header or to the base64 text body.

Modification:
Now, the bse64 text body always starts with a non-whitespace character, and any newlines in the body must follow after that.
This creates a fixture in the pattern that separates the header from the body unambiguously, and thus avoids exponential back-off.

Result:
More robust PEM parsing.