Explicitly include allocation in the LifecycleTracer API
Motivation:
The allocation is currently an implicit trace event, that happens as a side-effect of getting the tracer.
This is an asymmetry that makes it harder to extend the tracer API.

Modification:
Move the allocation trace point out of LifecycleTracer.get, and expose it as an explicit method on the API.
Then make sure it's called from the only place where we get a tracer.

Result:
Cleaner code