Allow to send custom events in both directions of the pipeline (#12497)
Motivation:

In the past it was only possible to send "custom events" in the inbound direction of the pipeline. There are some cases when it would also be useful to send these in the outbound direction. In this cases we used write(...) which is kind of odd and not consistent.

Modifications:

- Add triggerOutboundEvent(...) method and callbacks
- Rename userEventTriggered(...) to inboundEventTriggered(...) to make it more clear that its not only for user events and also to make it more consistent in terms of naming

Result:

Be able to let custom events flow in both directions of the pipeline