Avoid shadowing exceptions with dispose(obj)
Motivation:
Methods like Resource.dispose(obj) can throw exceptions, and are often used in finally-clauses.
When an exception is thrown from a finally-clause, it can shadow any exception that may have been thrown from the try-clause.

Modification:
Go through all of the places where Resource.dispose() is used in a finally clause, and change the code such that no exception from the try-clause will be shadowed if the dispose() method throws.
In many cases, this can be done by deferring to try-with-resources, but a small number of cases need special care.

Result:
We should no longer miss any root causes, when a Resource.dispose(obj) call throws.