Remove remaining references to ByteBuf
Motivation:
The adaptors will be moved to a contrib module, and we can drop our dependency on Netty 4.1.

Modification:
Remove the Netty 4.1 dependency, and all references to the classes it provided.
A lot of code, like tests and benchmarks, are ByteBuf specific and have been removed.
Some had not been fully converted, so this has been done.

Result:
We no longer need or use ByteBuf, or anything else from Netty 4.1.