`AbstractChannel.newId()` should be removed
Motivation:

We should better remove `AbstractChannel.newId()` so it can not be overriden as this method is called from within the constructor. If a sub-class of `AbstractChannel` wants to use its own `ChannelId` implementation it should just use the constructor of `AbstractChannel` that takes a ChannelId.

Modifications:

Remove `newId()` static

Result:

Cleanup code-smell