When name resolution is in progress ensure the Future is obtained from the correct Promise
Motivation:
When name resolution is in progress, the current Promise needs to wait for a notification from the previous Promise.
In the current implementation the Future is wrongly obtained from the current Promise instead of the previous Promise.

Modifications:

- `earlyFuture` is obtained from the `earlyPromise`

Result:

When name resolution is in progress, the Future is obtained from the correct Promise