Add Better Path Handling
Directly forward the uri when it's a path instead of trying to parse it.

Also clean up isOriginForm/isAsteriskForm as their logic seems to be incorrect (for example, URI#getSchemeSpecificPart() never returns null so they both return false trivially). They should probably be edited themselves in a later commit (and possibly used again here when correct). Since this was always returning true, it's unclear if authority logic is still correct after this change.