Fix mocking in AbstractChannelTest
Motivation:
The changes to how channels are registered with event loops has broken the mocking in this test.

Modification:
The event loop is now mocked to return completed futures for channel registration and deregistration, which is required to get their listeners to run, instead of throwing a NullPointerException.

Result:
AbstractChannelTest now passes again.