Fix Test Failures
Fix test failures. This requires refactoring the authority pathway to
reflect what it was in practice before (though not what the logic
implied).

Specifically, the logic before never set the HTTP2 Authority if the path
was in origin or asterisk form (although due to those methods always
returning false this never affected output).