Make sure all LocalAddresses are unique
Motivation:
As we run more tests in parallel, the risk of colliding local addresses increases, unless we make sure to give them unique names.

Modification:
Add a new constructor to LocalAddress which takes a class and generates a unique name with a UUID.
Use this constructor in place of fixed strings, and add randomness to the remaining dynamically named addresses.

Result:
We shouldn't see any more "address already in use" test failures.