Add BufferAllocatorMetric#pinnedMemory
Motivation:
The `used memory` is the amount of memory that a pooled allocator has currently allocated and committed for itself.
The `pinned memory` is the memory currently in use by buffers in circulation.
This is porting PR #11667 from `4.1` to `main`.

Modification:
- Add pinned memory accounting to PoolChunk.
- Add junit test

Result:
It is now possible to obtain the memory currently in use by buffers in circulation.