Add ProtocolEvent and let SSL and websockets event implement it
Motivation:

There are various protocols that will need to have a way to signal some sort of protocol event (like for example a protocol upgrade). We should have an interface for such events so its easier to "match" these.

Modifications:

- Add ProtocolEvent and implement it for SSL / websockets
- Adjust code

Result:

Easier to and more consist way of re-acting on protocol events