Expose new API EventLoop#supportsIoHandler
Motivation:

The `EventLoopGroup` can be
- `new MultithreadEventLoopGroup(NioHandler.newFactory())`
- `new MultithreadEventLoopGroup(EpollHandler.newFactory())`
- ...

A `Bootstrap` can be created like this
```
new Bootstrap().group(group)
               .channelFactory(eventLoop -> <new channel>)
```

The channel that will be returned by the channel factory may need to satisfy
- NioHandler#cast
- EpollHandler#cast
- ...

It will be helpful if one can check what IoHandler is supported by the event loop (if any is supported at all).

Modifications:

- Expose new API EventLoop#supportsIoHandler
- Add junit tests

Result:

It is possible to provide a channel based on the IoHandler supported by the event loop.