Replace InetAddressFamily with ProtocolFamily
Motivation:
Now that we base-line on JDK11 we can get rid of our own InetAddressFamily and just use ProtocolFamily / StandardProtocolFamily. This also opens up all kind of posibilities to support different protocol families in the future that are added to later JDK releases.

Modifications:

- Remove InetAddressFamily and related code
- Use ProtocolFamily as a replacement
- Add untility methods to NetUtil
- Adjust tests

Result:

Less code and more flexible