Correctly update pendingOutboundBytes() when handler is removed and also guard against invalid state when updating pendingOutboundBytes() (#12587)
Motivation:

We didnt update the pendingOutboundBytes() of the pipeline when a handler was removed so we could get out of sync. Also we didnt guard against invalid return values of `ChannelHandler.pendingOutboundBytes(...)` when update which could lead to side-effects.

Beside this we also should not try to call `channelExceptionCaught` when an IllegalStateException was thrown due of `pendingOutboundBytes(...)` as it would just throw another one as well

Modifications:

- Correctly update pending bytes on handler remoal
- Better handling of invalid return values
- Add unit tests

Result:

pendingOutboundBytes() are correctly updated in all cases.