timeout: expose remaining time of timeout
To allow timeouts to be a little more useful, expose the deadline of the
timeout via a method `remaining(TimeUnit)` which returns the remaining
time until the deadline in `TimeUnit`. Would possibly be useful to
instead of the `java.time.Instant` instead, but there's both a
difference in the underlying clock in most platforms (monotonic vs
realtime) and for some reason `java.time.Instant` doesn't pass the
animalsniffer check.