Add check for |Connection| and |Upgrade| headers inside WebSocketServerHandshaker13, see https://datatracker.ietf.org/doc/html/rfc6455#section-4.2.1 (#12606)
Motivation:

We faced with problem between proxies and Netty backend server. If we received a handshake request which doesn't contain a |Connection| or |Upgrade| headers with proper value the Netty accept this request and upgrade pipeline to handle websocket frames. But proxy thinks that is not upgrade request and can reuse this connection to any others http requests which we don't expect.

Modification:

Add check for |Connection| and |Upgrade| headers inside `WebSocketServerrHandshaker13` according spec
https://datatracker.ietf.org/doc/html/rfc6455#section-4.2.1. 

Result:

Proper handshake request handling for websocket version 13.