Change EventLoop / IoHandler to take an IoHandle for registration / deregistration (#12614)
Motivation:

By changing the EventLoop / IoHandler to take an IoHandle (and not a Channel) as argument we can support handling of other things then netty channels via the EventLoop / IoHandler. For example we can use it to let people register their self managed SelectableChannel implementations. In the future we might also use this abstraction to support file IO via io_uring.

Modifications:

- Add IoHandle interface
- Change EventLoop / IoHandler methods to take IoHandle
- Let Channel extend IoHandle
- Remove NioTask and replace it by NioSelectableChannelHandle
- Share processing logic in NioHandler.

Result:

More flexible and future proof API when it comes to IO handling