Netty 5 version of #12641
Motivation:

Before this patch, JdkZlibEncoder allocates two buffers of roughly the same size as the input buffer: one for copying data to heap, one for output. The former is always unnecessary, since inflate calls can take smaller buffers, and the latter is sometimes unnecessary, if the input data compresses well. This behavior can lead to large heap buffer allocation that bypass pooling and lead to unnecessary heap pressure.

Modification:

This patch introduces a limit to both allocations. The limit is arbitrarily set to 64K (should it be configurable?) to stay within standard pool thresholds.

The copy-to-heap code is modified so that if the input is larger than this limit, it is gradually passed to the deflater with multiple setInput/deflate calls.

The output buffer code was already capable of expanding a too small buffer. If the output is larger than the limit, we could instead write it incrementally as small chunks downstream, however I don't know if this is worthwhile, since we don't really have any way of handling backpressure.

Additionally, this patch fixes three bugs I found during development:

    It is possible for the deflater to return needsInput even when there is still data to write, because the output buffer is full. I changed the code to check whether output is full first. This could potentially lead to a pointless resize if we estimated the output size exactly right, but this seems unlikely.
    the size estimate in allocateBuffer could overflow if the input buffer is almost 2G. The capacity limit fixes this issue.
    The deflater kept a reference to the input array even after the encode call had completed. I set the input to new byte[0] to avoid this.

Result:

JdkZlibEncoder does not allocate unnecessarily large temporary arrays. However, it may have to resize the output array more often if data can't be compressed well.