Track independent refCnt in MixedAttribute and MixedFileUpload
Motivation:

Before this patch, MixedAttribute and MixedFileUpload delegated reference counting to their underlying memory or file representation. When the representation changes from memory to disk because the memory size limit was exceeded, the reference count of the old representation is not transferred, so the reference attribute resets to 1. This may leak the memory of the old representation, and may lead to exceptions when users try to release the mixed data.

Modification:

Introduce a new AbstractMixedHttpData class, analogous to the existing abstract classes for disk and memory data. This new class does its own reference counting, and unifies the swapping mechanism for attributes and file uploads.

The change is fairly big, but it's mostly copied from the old implementations. Makes it a bit annoying to review though, so I might have missed a bug.

I tried inheriting from AbstractHttpData, but it didn't really reduce the code size much, and makes some code less clear because it leads to two copies of the name, charset, etc.

Regarding compatibility, I managed to fix most compat issues reported by revapi. However inheriting from a new class in a non-final class is apparently a breaking change no matter what we do, and additionally AbstractReferenceCounted.setRefCnt is problematic. I think both of these errors are unlikely to affect anyone, and are justifiable. I've added an exclusion to the pom.

There is a new test class to test the retain/refCnt behavior. I didn't add it to the HttpDataTest, because the data defined there does not have a size limit we can use to test easily.

Result:

MixedAttribute and MixedFileUpload now behave like any other reference counted object.