Avoid including header values in header validation failure exceptions (#12642)
Motivation:
Exception messages may end up in logs, or other error reporting mechanisms, and we cannot necessarily tell at header validation time if the header value is sensitive information or not.
Hence, it's better to not include it, and instead rely on the header name as a debugging aid.

Modification:
Remove the inclusion of the header value that fails header validation.
And instead add the name of the header that fails validation.

Result:
If sensitive information fails header validation, it will no longer be exposed through the exception messages.