Add Blockhound exceptions for the PooledByteBufAllocator (#12653)
Motivation:
In #12622 we replaced a number of synchronised blocks with explicit locks.
It turns out that blockhound ignores synchronised blocks for some reason, and now complain about the explicit locks.

Modification:
Add exceptions and allow blocking calls in select methods in the PooledByteBufAllocator.
Also add a test that create races inside the pooled allocator in an attempt to get the thread blocked on the locks while blockhound is watching.

Result:
No more blockhound complaints.