Remove shutdown hooks for default buffer allocators
Motivation:
Shutdown hooks can race with application threads, and this can cause problems since the allocators may still be in use during the shutdown process.

It is also not necessary to shut the allocators down, since the process is about to terminate anyway, at which point all memory will be returned to the OS regardlesss of what we do.

Modification:
Avoid installing any shutdown hooks for closing the DefaultBufferAllocators.

Result:
We should no longer see any "closed allocator exceptions" when shutting the JVM down - at least not any that are caused by these shared allocators.

Fixes #12671