Handle content-type optional parameters other than charset (netty#12663) (#12664)
Motivation:

When receiving multipart/form-data where the form-data includes a file with optional parameter, in my case a codec parameter, the content-type was incorrectly set to the parameter value. Currently this only behaves correctly if the optional parameter is a character set parameter.

Modification:

When processing the content-type header contents, after splitting the header into the String[], check each item to see if it is a name=value pair, indicating it is an optional parameter. Add an attribute using the split name and value.

Result:

When processing the content-type header, the code now correctly sets the content-type and adds additional attributes for the optional parameters.

Fixes #12663