Move logic for read handling into AbstractChannel
Motivation:

There was a lot of duplication in the different Channel implementation when it comes to handling reads. To make things more consistent and also easier for transport implementators we should move all the complexity into AbstractChannel.

Modifications:

- Introduce AbstractChannel.readNow(...) which should be called when there is data to read from the underlying transport (like a socket).
- Add writeFlushedNow() which should be called when the underlying transport becomes writable again
- Move readPending tracking and autoRead logic into AbstractChannel
- Thighten up access and visibility
- Remove a lot of code duplication

Result:

Cleaner and more consistent code. Beside this also easier to write a custom Channel / Transport implementation