Merge readable and writable buffer component iteration
Motivation:
It should be possible to iterate the components of a buffer, even when it has no readable or writable bytes.
For the purpose of pointer arithmetic, it should also be possible to get a base pointer for off-heap buffers.
It additionally turns out that unifying ReadableComponent and WritableComponent into a single BufferComponent, allow us to make the iteration entirely allocation free for non-composite buffers.
This in turn allow us to simplify the API by removing the internal iteration methods and only keep a single forEachComponent iteration method.

Modification:
- Remove WritableComponent and ReadableComponent, and their processors.
- And in their place have a single BufferComponent which contain both of the APIs from the readable and writable components.
- Add a baseNativeAddress method to BufferComponent.
- Change all usages sites of both the internal and external versions of forEachReadable/Writable, to use the new forEachComponent.
- Add firstReadable and firstWritable convenience methods to ComponentIterator, which will skip over components that are not readable or writable, respectively.
- Add nextReadable and nextWritable methods to ComponentIterator.Next, which do the same.
- Change places like EngineWrapper, IovArray, etc. to no longer rely on internal iteration but use external iteration instead.
- Remove processor classes that only existed to be used for internal iteration.
- Read-only buffers no longer have any "writable bytes".

Result:
Iterating the components of non-composite buffers is now always allocation-free; no capturing-lambda or iterator objects are constructed.
The APIs are also simpler; fewer methods and fewer classes.

Fixes #12609 and closes #12611