UnsafeBuffer don't need to zero memory on allocation
Motivation:
Our Buffer API (and the ByteBuf API before it) do not guarantee that newly allocated buffers are zeroed.
Our pooling buffer allocator already do not zero buffers when reusing them.
I discovered that our UnsafeMemoryManager did zeroing, and then I also discovered that Intellij did not properly activate the unsafeBuffer profile; it activated but had no effect.

Modification:
Remove zeroing from the UnsafeMemoryManager, when it allocates a new buffer.
Fix all of the tests that started failing.
Change the effect of activating the unsafeBuffer profile, so that it works around the different property resolution in Intellij.

Result:
Allocating unsafe, off-heap buffers is now faster, when not using pooling.
And the unsafeBuffer profile now has the effect you would expect, when enabled in Intellij.