Improve logging when MacOSDnsServerAddressStreamProvider cannot be found/loaded
Motivation:

When `MacOSDnsServerAddressStreamProvider` cannot be found/loaded, a warning/error appears
in the logs, specifying that without this functionality the DNS resolution might be incorrect.
However there is no guidance how this can be fixed.

Modifications:

- Extend the log message with information for the dependency that is needed in order to resolve this issue.

Result:
Users will be informed for the missing dependency that will resolve the issue.
Related to #11020