Fix bug in SslContext private key reading fall-back path
Motivation:
We first try to read private keys with BouncyCastle, and then fall back to our own code if that fails.
It could be the case, that BouncyCastle read some data before it failed, and in that case we need to make sure that our fallback code gets to read the key data from the original stream position - not whatever position caused a failure in BouncyCastle.

Modification:
Use the mark and reset methods of InputStream to reset the key stream position, if BouncyCastle cannot read a key.
If the provided input stream does not already support marking, then we wrap it in a BufferedInputStream which does.
We allow up to 1 MiB of data to be buffered for this purpose.
If the key is bigger than this, and BouncyCastle reads more than this before failing, then resetting will throw an exception so we don't get silent failures.

Result:
Fixed private key fallback code path.
Fixes #12745