Fix leak detection around const buffer suppliers
Motivation:
Const-buffer suppliers have no direct way to close them, so they can only be closed by `Cleaners`.
We should not consider this a leak.

Modification:
Make sure that the parent buffer used in a const-buffer supplier always gets a `CleanerDrop`, and that this drop will not report it to the `LeakDetection`.
Update tests to cover this particular case.
Make the `LoggingLeakCallback` only be used for reporting when there are no other leak callbacks installed - otherwise tests that intentionally cause leaks will fail the `-Pleak` build.
Fix a number of leaks in tests.

Result:
No more leaks reported by the use of `BufferAllocator#constBufferSupplier`.