Add a 5-minute timeout to CleanerDropTest
Motivation:
It appears this test occasionally fails on Java 18.
When it does, it currently manifests as the build getting stuck for hours.
Unfortunately, we do need to check that all leaked objects are captured by the installed leak detector, as the build will otherwise fail due to leaks being detected.

Modification:
Add a timeout that fails the test after 5 minutes, instead of waiting for the entire build to time out.

Result:
This test will now fail faster, when it fails.