Make PoolSubpage more resilient to certain errors
Motivation:
We occasionally see a strange `ArrayIndexOutOfBoundsException: Index 67108863 out of bounds for length 8` exception, which is caused by `getNextAvail()` returning `-1`, even though `numAvail` is positive.

Modification:
Improve the error message when this occurs, and include a few more details to verify the assumption that `bitmapIdx` really is `-1` even though `numAvail` is positive.
Also remove the PoolSubpage from the arena when this happens, as it likely means the subpage is in a broken state, and we should not use it anymore.

Result:
We have no fixed the root cause, but hopefully the system will be better able to recover, or at least we'll get slightly better diagnostics.