WebSocket13FrameEncoder should override encode instead of encodeAndClose (#12780)
Motivation:

The #12773 PR has fixed a memory leak from the WebSocket13FrameEncoder.encodeAndClose method, where the websocket frame to be encoded was not always closed properly.

Now, @amizurov has suggested to do a better fix by just letting the WebSocket13FrameEncoder override the _encode_ method instead of the _encodeAndClose_. It give an autoclose for the input argument and don't need to do this manually, behaviour will be similar with 4.1.

However, care must be taken when the frame is itself appended to the "out" list of produced data. In 4.x, this was done using "data.retain", with the new Buffer API, the data can be added to the "out" list using "data.split".

Modification:

The WebSocket13FrameEncoder is now overriding the _encode_ method instead of the _encodeAndClose_, and the frame message is not closed anymore at all (this is done by the _encodeAndClose_ method from the superclass).
And when the frame message data has to be appended into the "out" list of produced encoded data, then it is done using "data.split" 

Result:

The WebSocket13FrameEncoder class is simpler and don't have to close the frame data parameter.