Remove `api` from the `io.netty5.buffer.api` package path
Motivation:
The `api` component is a holdover from before the package rename from `io.netty` to `io.netty5`, and from before `ByteBuf` was removed.
So this was originally meant to separate the new code from the old.
But since we did the package rename, and we removed `ByteBuf`, this is no longer necessary.

Modification:
Remove `api` from the package path and update... everything... to match.

Result:
Buffers are now in a package path more in line with the naming convention in the rest of Netty.